decoder: recurse structures, not form values (#8)
* decoder: recurse structures, not form values

Decoding using purely form values can leave existing values in place for certain types. The existing solution in 3a844d66a01c4270d9ccdea54bce3ed59259b617 was an improvement to this scenario, but pre-filling all data points with empty keys could lead to slices with empty strings in them, when the intended result was a slice with no elements. The emptySlice test in decoder_test.go verifies that this issue has been resolved.

* decoder: add show conditions to HTTPDecoder.

Show Conditions are now required in the decoder so that hidden fields are not decoded into when they were not present in the form.